working on GNSSController
